split out service work into a service file
